+++ src/utils/math.ts
