+++ src/utils/metadata.ts
@@ -1,6 +1,7 @@
 import { FileMetadata } from "./customTypes";
-import * as isi from "./isiUtils";
-import * as scopus from "./scopusUtils";
+import { mostCommon } from "./arrays";
+import * as isi from "./isi";
+import * as scopus from "./scopus";
 import md5 from "md5";
 
 const metadata = async (name: string, blob: Blob): Promise<FileMetadata> => {
@@ -12,7 +13,7 @@ const metadata = async (name: string, blob: Blob): Promise<FileMetadata> => {
       name,
       blob,
       hash,
-      keywords: isi.mostCommon(isi.keywords(content), 3),
+      keywords: mostCommon(isi.keywords(content), 3),
       articles: isi.countArticles(content),
       citations: isi.countReferences(content),
       valid: true,
@@ -24,7 +25,7 @@ const metadata = async (name: string, blob: Blob): Promise<FileMetadata> => {
       name,
       blob,
       hash,
-      keywords: isi.mostCommon(scopus.keywords(content), 3),
+      keywords: mostCommon(scopus.keywords(content), 3),
       articles: scopus.countArticles(content),
       citations: scopus.countReferences(content),
       valid: true,
